requirements.txt
yt_dlp
python-vlc